Allow empty keys in project settings
This change means that there is a strict hierarchy of settings:
- defaults
- user settings
- project settings

At any level a null (including empty) value will now use the value
from the layer above. Previously null values in the project level
used the default value of the key, which was surprising.

Note: this is could be a breaking change for projects, but I believe
it is acceptable because:
- most projects with custom settings will likely override everything,
  even if the values are the same as the defaults
- the current behaviour for this is undocumented
- the new behaviour more closely matches how vscode's own preferences
  behave (and I believe is therefore what most users would expect)